\name{IPMRF-internal}
\alias{prevtree}
\alias{prevtreeparty}
\alias{number2binary}
\title{Internal function(s)}
\description{They are internal functions used for returning which variables are used in prediction in tree with \pkg{randomForest} and \pkg{party}. }
\author{Irene Epifanio}
\keyword{internal}

