% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipmrf-package.R
\docType{package}
\name{IPMRF-package}
\alias{IPMRF-package}
\title{IPMRF: Intervention in Prediction Measure (IPM) for Random Forests}
\description{
It computes IPM for assessing variable importance for random forests. See I. Epifanio (2017).
Intervention in prediction measure: a new approach to assessing variable importance for random
forests. BMC Bioinformatics.
}
\section{IPMRF functions}{

\itemize{
\item \code{\link{ipmparty}}: IPM casewise with CIT-RF by \pkg{party} for OOB samples
\item \code{\link{ipmpartynew}}: IPM casewise with CIT-RF by \pkg{party} for new samples
\item \code{\link{ipmrf}}: IPM casewise with CART-RF by \pkg{randomForest} for OOB samples
\item \code{\link{ipmrfnew}}: IPM casewise with CART-RF by \pkg{randomForest} for new samples
\item \code{\link{ipmranger}}: IPM casewise with RF by \pkg{ranger} for OOB samples
\item \code{\link{ipmrangernew}}: IPM casewise with RF by \pkg{ranger} for new samples
\item \code{\link{ipmgbmnew}}: IPM casewise with GBM by \pkg{gbm} for new samples
}
}

\references{
Pierola, A. and Epifanio, I. and Alemany, S. (2016) An ensemble of ordered logistic regression
and random forest for child garment size matching. \emph{Computers & Industrial Engineering},
\bold{101}, 455--465.

Epifanio, I. (2017) Intervention in prediction measure: a new approach to assessing variable
importance for random forests. \emph{BMC Bioinformatics}, \bold{18}, 230.
}
\seealso{
https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-017-1650-8
}
\author{
Irene Epifanio, Aleix Alcacer, Stefano Nembrini
}
